@@ -18,7 +18,7 @@ resolver = "2"
 
 [package]
 name = "kanata"
-version = "1.8.0"
+version = "1.9.0"
 authors = ["jtroo <j.andreitabs@gmail.com>"]
 description = "Multi-layer keyboard customization"
 keywords = ["keyboard", "layout", "remapping"]
@@ -27,7 +27,7 @@ homepage = "https://github.com/jtroo/kanata"
 repository = "https://github.com/jtroo/kanata"
 readme = "README.md"
 license = "LGPL-3.0-only"
-edition = "2021"
+edition = "2024"
 default-run = "kanata"
 
 [lib]
@@ -43,9 +43,9 @@ anyhow = "1"
 clap = { version = "4", features = [ "std", "derive", "help", "suggestions" ], default-features = false }
 dirs = "5.0.1"
 indoc = { version = "2.0.4", optional = true }
-instant = { version = "0.1.12", default-features = false }
 log = { version = "0.4.8", default-features = false }
 miette = { version = "5.7.0", features = ["fancy"] }
+notify-debouncer-mini = { version = "0.4", optional = true }
 once_cell = "1"
 parking_lot = "0.12"
 radix_trie = "0.2"
@@ -53,21 +53,17 @@ rustc-hash = "1.1.0"
 simplelog = "0.12.0"
 serde_json = { version = "1", features = ["std"], default-features = false, optional = true }
 time = "0.3.36"
-# kanata-keyberon = "0.180.1"
-# kanata-parser = "0.180.1"
-# kanata-tcp-protocol = "0.180.1"
-# Uncomment below and comment out above for testing local changes.
-# Otherwise any changes to the local files will not reflect in the compiled
-# binary.
-kanata-keyberon = { path = "keyberon" }
-kanata-parser =   { path = "parser" }
-kanata-tcp-protocol = { path = "tcp_protocol" }
+web-time = "1.1.0"
+
+kanata-keyberon = { path = "keyberon", version = "0.190.0" }
+kanata-parser =   { path = "parser", version = "0.190.0" }
+kanata-tcp-protocol = { path = "tcp_protocol", version = "0.190.0" }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 arboard = "3.4"
 
 [target.'cfg(target_os = "macos")'.dependencies]
-karabiner-driverkit = "0.1.4"
+karabiner-driverkit = "0.1.5"
 objc = "0.2.7"
 core-graphics = "0.24.0"
 open = { version = "5", optional = true }
@@ -75,7 +71,7 @@ libc = "0.2"
 os_pipe = "1.2.1"
 
 [target.'cfg(target_os = "linux")'.dependencies]
-evdev = "0.12.2"
+evdev = "0.13.0"
 inotify = { version = "0.10.0", default-features = false }
 mio = { version = "0.8.11", features = ["os-poll", "os-ext"] }
 nix = { version = "0.26.1", features = ["ioctl"] }
@@ -89,6 +85,9 @@ winapi = { version = "0.3.9", features = [
     "wincon",
     "timeapi",
     "mmsystem",
+    "winuser",
+    "windef",
+    "minwindef",
 ] }
 windows-sys = { version = "0.52.0", features = [
     "Win32_Devices_DeviceAndDriverInstallation",
@@ -122,7 +121,7 @@ indoc = { version = "2.0.4", optional = true }
 regex = { version = "1.10.4", optional = true }
 
 [features]
-default = ["tcp_server","win_sendinput_send_scancodes", "zippychord"]
+default = ["tcp_server","win_sendinput_send_scancodes", "zippychord", "watch"]
 perf_logging = []
 tcp_server = ["serde_json"]
 win_sendinput_send_scancodes = ["kanata-parser/win_sendinput_send_scancodes"]
@@ -133,7 +132,6 @@ interception_driver = ["kanata-interception", "kanata-parser/interception_driver
 simulated_output = ["indoc"]
 simulated_input = ["indoc"]
 passthru_ahk = ["simulated_input","simulated_output"]
-wasm = [ "instant/wasm-bindgen" ]
 gui = ["win_manifest","kanata-parser/gui",
   "win_sendinput_send_scancodes","win_llhook_read_scancodes",
   "muldiv","strip-ansi-escapes","open",
@@ -142,6 +140,7 @@ gui = ["win_manifest","kanata-parser/gui",
   "native-windows-gui/tray-notification","native-windows-gui/message-window","native-windows-gui/menu","native-windows-gui/cursor","native-windows-gui/high-dpi","native-windows-gui/embed-resource","native-windows-gui/image-decoder","native-windows-gui/notice","native-windows-gui/animation-timer",
 ]
 zippychord = ["kanata-parser/zippychord"]
+watch = ["notify-debouncer-mini"]
 
 [profile.release]
 opt-level = "z"
