@@ -1,8 +1,10 @@
-use anyhow::{anyhow, Error, Result};
+use anyhow::{Error, Result, anyhow};
 use std::net::SocketAddr;
 use std::path::PathBuf;
 use std::str::FromStr;
 
+#[cfg(feature = "watch")]
+pub mod file_watcher;
 #[cfg(all(target_os = "windows", feature = "gui"))]
 pub mod gui;
 pub mod kanata;
@@ -23,6 +25,8 @@ pub struct ValidatedArgs {
     #[cfg(target_os = "linux")]
     pub symlink_path: Option<String>,
     pub nodelay: bool,
+    #[cfg(feature = "watch")]
+    pub watch: bool,
 }
 
 pub fn default_cfg() -> Vec<PathBuf> {
@@ -52,7 +56,7 @@ impl FromStr for SocketAddrWrapper {
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         let mut address = s.to_string();
         if let Ok(port) = s.parse::<u16>() {
-            address = format!("127.0.0.1:{}", port);
+            address = format!("127.0.0.1:{port}");
         }
         address
             .parse::<SocketAddr>()
