@@ -44,8 +44,7 @@ pub(crate) fn parse_on_idle_fakekey(
     ac_params: &[SExpr],
     s: &ParserState,
 ) -> Result<&'static KanataAction> {
-    const ERR_MSG: &str =
-        "on-idle-fakekey expects three parameters:\n<fake key name> <(tap|press|release)> <idle time>\n";
+    const ERR_MSG: &str = "on-idle-fakekey expects three parameters:\n<fake key name> <(tap|press|release)> <idle time>\n";
     if ac_params.len() != 3 {
         bail!("{ERR_MSG}");
     }
@@ -165,8 +164,7 @@ fn parse_delay(
     is_release: bool,
     s: &ParserState,
 ) -> Result<&'static KanataAction> {
-    const ERR_MSG: &str = "fakekey-delay expects a single number (ms, 0-65535)";
-    log::warn!("The configuration contains a fakekey-delay action. This is broken for many use cases. It is recommended to use macro instead.");
+    const ERR_MSG: &str = "delay expects a single number (ms, 0-65535)";
     let delay = ac_params[0]
         .atom(s.vars())
         .map(str::parse::<u16>)
@@ -257,6 +255,28 @@ pub(crate) fn parse_on_idle(ac_params: &[SExpr], s: &ParserState) -> Result<&'st
 
     Ok(s.a.sref(Action::Custom(s.a.sref(s.a.sref_slice(
         CustomAction::FakeKeyOnIdle(FakeKeyOnIdle {
+            coord,
+            action,
+            idle_duration,
+        }),
+    )))))
+}
+
+pub(crate) fn parse_on_physical_idle(
+    ac_params: &[SExpr],
+    s: &ParserState,
+) -> Result<&'static KanataAction> {
+    const ERR_MSG: &str =
+        "on-physical-idle expects three parameters: <timeout> <action> <key-name>";
+    if ac_params.len() != 3 {
+        bail!("{ERR_MSG}");
+    }
+    let idle_duration = parse_non_zero_u16(&ac_params[0], s, "on-idle-timeout")?;
+    let action = parse_vkey_action(&ac_params[1], s)?;
+    let coord = parse_vkey_coord(&ac_params[2], s)?;
+
+    Ok(s.a.sref(Action::Custom(s.a.sref(s.a.sref_slice(
+        CustomAction::FakeKeyOnPhysicalIdle(FakeKeyOnIdle {
             coord,
             action,
             idle_duration,
