@@ -3,13 +3,13 @@
     allow(dead_code, unused_imports, unused_variables, unused_mut)
 )]
 
-use anyhow::{anyhow, bail, Result};
-use evdev::{InputEvent, InputEventKind, RelativeAxisType};
+use anyhow::{Result, anyhow, bail};
+use evdev::{EventSummary, InputEvent, RelativeAxisCode};
 use log::info;
 use parking_lot::Mutex;
 use std::convert::TryFrom;
-use std::sync::mpsc::SyncSender as Sender;
 use std::sync::Arc;
+use std::sync::mpsc::SyncSender as Sender;
 
 use super::*;
 
@@ -21,6 +21,7 @@ impl Kanata {
 
         let k = kanata.lock();
         let allow_hardware_repeat = k.allow_hardware_repeat;
+        let mouse_movement_key = k.mouse_movement_key.clone();
         let mut kbd_in = match KbdIn::new(
             &k.kbd_in_paths,
             k.continue_if_no_devices,
@@ -44,6 +45,15 @@ impl Kanata {
             log::trace!("event count: {}\nevents:\n{events:?}", events.len());
 
             for in_event in events.iter().copied() {
+                if let Some(ms_mvmt_key) = *mouse_movement_key.lock() {
+                    if let EventSummary::RelativeAxis(_, _, _) = in_event.destructure() {
+                        let fake_event = KeyEvent::new(ms_mvmt_key, KeyValue::Tap);
+                        if let Err(e) = tx.try_send(fake_event) {
+                            bail!("failed to send on channel: {}", e)
+                        }
+                    }
+                }
+
                 let key_event = match KeyEvent::try_from(in_event) {
                     Ok(ev) => ev,
                     _ => {
@@ -69,20 +79,30 @@ impl Kanata {
                     if !handle_scroll(&kanata, in_event, key_event.code, &events)? {
                         continue;
                     }
-                } else {
-                    // Handle normal keypresses.
-                    // Check if this keycode is mapped in the configuration.
-                    // If it hasn't been mapped, send it immediately.
-                    if !MAPPED_KEYS.lock().contains(&key_event.code) {
-                        let mut kanata = kanata.lock();
-                        #[cfg(not(feature = "simulated_output"))]
-                        kanata
-                            .kbd_out
-                            .write_raw(in_event)
-                            .map_err(|e| anyhow!("failed write: {}", e))?;
-                        continue;
-                    };
                 }
+
+                match key_event.value {
+                    KeyValue::Release => {
+                        PRESSED_KEYS.lock().remove(&key_event.code);
+                    }
+                    KeyValue::Press => {
+                        PRESSED_KEYS.lock().insert(key_event.code);
+                    }
+                    _ => {}
+                }
+
+                // Handle normal keypresses.
+                // Check if this keycode is mapped in the configuration.
+                // If it hasn't been mapped, send it immediately.
+                if !MAPPED_KEYS.lock().contains(&key_event.code) {
+                    let mut kanata = kanata.lock();
+                    #[cfg(not(feature = "simulated_output"))]
+                    kanata
+                        .kbd_out
+                        .write_raw(in_event)
+                        .map_err(|e| anyhow!("failed write: {}", e))?;
+                    continue;
+                };
 
                 // Send key events to the processing loop
                 if let Err(e) = tx.try_send(key_event) {
@@ -138,10 +158,10 @@ fn handle_scroll(
 ) -> Result<bool> {
     let direction: MWheelDirection = code.try_into().unwrap();
     let scroll_distance = in_event.value().unsigned_abs() as u16;
-    match in_event.kind() {
-        InputEventKind::RelAxis(axis_type) => {
+    match in_event.destructure() {
+        EventSummary::RelativeAxis(_, axis_type, _) => {
             match axis_type {
-                RelativeAxisType::REL_WHEEL | RelativeAxisType::REL_HWHEEL => {
+                RelativeAxisCode::REL_WHEEL | RelativeAxisCode::REL_HWHEEL => {
                     if MAPPED_KEYS.lock().contains(&code) {
                         return Ok(true);
                     }
@@ -158,10 +178,12 @@ fn handle_scroll(
                     let mut kanata = kanata.lock();
                     if !all_events.iter().any(|ev| {
                         matches!(
-                            ev.kind(),
-                            InputEventKind::RelAxis(
-                                RelativeAxisType::REL_WHEEL_HI_RES
-                                    | RelativeAxisType::REL_HWHEEL_HI_RES
+                            ev.destructure(),
+                            EventSummary::RelativeAxis(
+                                _,
+                                RelativeAxisCode::REL_WHEEL_HI_RES
+                                    | RelativeAxisCode::REL_HWHEEL_HI_RES,
+                                _
                             )
                         )
                     }) {
@@ -172,7 +194,7 @@ fn handle_scroll(
                     }
                     Ok(false)
                 }
-                RelativeAxisType::REL_WHEEL_HI_RES | RelativeAxisType::REL_HWHEEL_HI_RES => {
+                RelativeAxisCode::REL_WHEEL_HI_RES | RelativeAxisCode::REL_HWHEEL_HI_RES => {
                     if !MAPPED_KEYS.lock().contains(&code) {
                         // Passthrough if the scroll wheel event is not mapped
                         // in the configuration.
