@@ -11,9 +11,6 @@ mod interception;
 #[cfg(all(not(feature = "simulated_input"), not(feature = "interception_driver")))]
 mod llhook;
 
-pub static PRESSED_KEYS: Lazy<Mutex<HashSet<OsCode>>> =
-    Lazy::new(|| Mutex::new(HashSet::default()));
-
 pub static ALTGR_BEHAVIOUR: Lazy<Mutex<AltGrBehaviour>> =
     Lazy::new(|| Mutex::new(AltGrBehaviour::default()));
 
@@ -154,7 +151,11 @@ impl Kanata {
                 log::info!("Requested live reload of file: {}", path.display(),);
             }
             None => {
-                log::error!("Requested live reload of config file number {}, but only {} config files were passed", n+1, self.cfg_paths.len());
+                log::error!(
+                    "Requested live reload of config file number {}, but only {} config files were passed",
+                    n + 1,
+                    self.cfg_paths.len()
+                );
             }
         }
         // if let Err(e) = self.do_live_reload(&None) {
@@ -163,5 +164,37 @@ impl Kanata {
         // }
         self.do_live_reload(&None)?;
         Ok(())
+    }
+}
+
+/// If kanata has been inactive for long enough, clear all states.
+/// This won't trigger if there are macros running, or if a key is
+/// held down for a long time and is sending OS repeats. The reason
+/// for this code is in case like Win+L which locks the Windows
+/// desktop. When this happens, the Win key and L key will be stuck
+/// as pressed in the kanata state because LLHOOK kanata cannot read
+/// keys in the lock screen or administrator applications. So this
+/// is heuristic to detect such an issue and clear states assuming
+/// that's what happened.
+///
+/// Only states in the normal key row are cleared, since those are
+/// the states that might be stuck. A real use case might be to have
+/// a fake key pressed for a long period of time, so make sure those
+/// are not cleared.
+#[cfg(all(not(feature = "interception_driver"), target_os = "windows"))]
+pub fn clear_states_from_inactivity(
+    k: &mut parking_lot::MutexGuard<Kanata>,
+    now: web_time::Instant,
+    last_input_time: web_time::Instant,
+    idle_clear_happened: &mut bool,
+) {
+    if (now - (last_input_time)) > time::Duration::from_secs(LLHOOK_IDLE_TIME_SECS_CLEAR_INPUTS)
+        && !*idle_clear_happened
+    {
+        *idle_clear_happened = true;
+        log::debug!("clearing keyberon normal key states due to inactivity");
+        let layout = k.layout.bm();
+        release_normalkey_states(layout);
+        PRESSED_KEYS.lock().clear();
     }
 }
