@@ -2,7 +2,7 @@
 // disable default console for a Windows GUI app
 mod main_lib;
 
-use anyhow::{bail, Result};
+use anyhow::{Result, bail};
 use clap::Parser;
 use kanata_parser::cfg;
 use kanata_state_machine::*;
@@ -42,6 +42,10 @@ kanata.kbd in the current working directory and
     #[arg(short, long, verbatim_doc_comment)]
     cfg: Option<Vec<PathBuf>>,
 
+    /// Read configuration from stdin instead of a file.
+    #[arg(long, verbatim_doc_comment)]
+    cfg_stdin: bool,
+
     /// Port or full address (IP:PORT) to run the optional TCP server on. If blank,
     /// no TCP port will be listened on.
     #[cfg(feature = "tcp_server")]
@@ -52,6 +56,7 @@ kanata.kbd in the current working directory and
         verbatim_doc_comment
     )]
     tcp_server_address: Option<SocketAddrWrapper>,
+
     /// Path for the symlink pointing to the newly-created device. If blank, no
     /// symlink will be created.
     #[cfg(target_os = "linux")]
@@ -59,7 +64,11 @@ kanata.kbd in the current working directory and
     symlink_path: Option<String>,
 
     /// List the keyboards available for grabbing and exit.
-    #[cfg(target_os = "macos")]
+    #[cfg(any(
+        target_os = "macos",
+        target_os = "linux",
+        all(target_os = "windows", feature = "interception_driver")
+    ))]
     #[arg(short, long)]
     list: bool,
 
@@ -98,6 +107,10 @@ kanata.kbd in the current working directory and
     /// configuration but want to default to no logging.
     #[arg(long, verbatim_doc_comment)]
     log_layer_changes: bool,
+
+    /// Watch configuration files for changes and reload automatically
+    #[arg(long, verbatim_doc_comment)]
+    watch: bool,
 }
 
 #[cfg(not(feature = "gui"))]
@@ -105,16 +118,41 @@ mod cli {
     use super::*;
 
     /// Parse CLI arguments and initialize logging.
-    fn cli_init() -> Result<ValidatedArgs> {
+    fn cli_init() -> Result<(ValidatedArgs, Option<String>)> {
         let args = Args::parse();
 
         #[cfg(target_os = "macos")]
         if args.list {
-            karabiner_driverkit::list_keyboards();
+            main_lib::list_devices_macos();
             std::process::exit(0);
         }
 
-        let cfg_paths = args.cfg.unwrap_or_else(default_cfg);
+        #[cfg(target_os = "linux")]
+        if args.list {
+            main_lib::list_devices_linux();
+            std::process::exit(0);
+        }
+
+        #[cfg(all(target_os = "windows", feature = "interception_driver"))]
+        if args.list {
+            main_lib::list_devices_windows();
+            std::process::exit(0);
+        }
+
+        let config_string = if args.cfg_stdin {
+            use std::io::Read;
+            let mut buf = String::new();
+            std::io::stdin().read_to_string(&mut buf)?;
+            Some(buf)
+        } else {
+            None
+        };
+
+        let cfg_paths = if config_string.is_none() {
+            args.cfg.unwrap_or_else(default_cfg)
+        } else {
+            vec![]
+        };
 
         let log_lvl = match (args.debug, args.trace, args.quiet) {
             (_, true, false) => LevelFilter::Trace,
@@ -145,24 +183,37 @@ mod cli {
         #[cfg(all(feature = "interception_driver", target_os = "windows"))]
         log::info!("using the Interception driver for keyboard IO");
 
-        if let Some(config_file) = cfg_paths.first() {
-            if !config_file.exists() {
-                bail!(
-                "Could not find the config file ({})\nFor more info, pass the `-h` or `--help` flags.",
-                cfg_paths[0].to_str().unwrap_or("?")
-            )
+        if config_string.is_none() {
+            if let Some(config_file) = cfg_paths.first() {
+                if !config_file.exists() {
+                    bail!(
+                        "Could not find the config file ({})\nFor more info, pass the `-h` or `--help` flags.",
+                        cfg_paths[0].to_str().unwrap_or("?")
+                    )
+                }
+            } else {
+                bail!("No config files provided\nFor more info, pass the `-h` or `--help` flags.");
             }
-        } else {
-            bail!("No config files provided\nFor more info, pass the `-h` or `--help` flags.");
         }
 
         if args.check {
             log::info!("validating config only and exiting");
-            let status = match cfg::new_from_file(&cfg_paths[0]) {
-                Ok(_) => 0,
-                Err(e) => {
-                    log::error!("{e:?}");
-                    1
+            let status = if let Some(ref cfg_str) = config_string {
+                use rustc_hash::FxHashMap;
+                match cfg::new_from_str(cfg_str, FxHashMap::default()) {
+                    Ok(_) => 0,
+                    Err(e) => {
+                        log::error!("{e:?}");
+                        1
+                    }
+                }
+            } else {
+                match cfg::new_from_file(&cfg_paths[0]) {
+                    Ok(_) => 0,
+                    Err(e) => {
+                        log::error!("{e:?}");
+                        1
+                    }
                 }
             };
             std::process::exit(status);
@@ -179,22 +230,36 @@ mod cli {
             cfg_forced::force_log_layer_changes(true);
         }
 
-        Ok(ValidatedArgs {
-            paths: cfg_paths,
-            #[cfg(feature = "tcp_server")]
-            tcp_server_address: args.tcp_server_address,
-            #[cfg(target_os = "linux")]
-            symlink_path: args.symlink_path,
-            nodelay: args.nodelay,
-        })
+        Ok((
+            ValidatedArgs {
+                paths: cfg_paths,
+                #[cfg(feature = "tcp_server")]
+                tcp_server_address: args.tcp_server_address,
+                #[cfg(target_os = "linux")]
+                symlink_path: args.symlink_path,
+                nodelay: args.nodelay,
+                #[cfg(feature = "watch")]
+                watch: args.watch,
+            },
+            config_string,
+        ))
     }
 
     pub(crate) fn main_impl() -> Result<()> {
-        let args = cli_init()?;
-        let kanata_arc = Kanata::new_arc(&args)?;
+        let (args, config_string) = cli_init()?;
+
+        let kanata_arc = if let Some(cfg_str) = config_string {
+            use rustc_hash::FxHashMap;
+            let kanata = Kanata::new_from_str(&cfg_str, FxHashMap::default())?;
+            std::sync::Arc::new(parking_lot::Mutex::new(kanata))
+        } else {
+            Kanata::new_arc(&args)?
+        };
 
         if !args.nodelay {
-            log::info!("Sleeping for 2s. Please release all keys and don't press additional ones. Run kanata with --help to see how understand more and how to disable this sleep.");
+            log::info!(
+                "Sleeping for 2s. Please release all keys and don't press additional ones. Run kanata with --help to see how understand more and how to disable this sleep."
+            );
             std::thread::sleep(std::time::Duration::from_secs(2));
         }
 
@@ -229,6 +294,14 @@ mod cli {
         if let (Some(server), Some(nrx)) = (server, nrx) {
             #[allow(clippy::unit_arg)]
             Kanata::start_notification_loop(nrx, server.connections);
+        }
+
+        // Start comprehensive file watcher if enabled (supports include files)
+        #[cfg(feature = "watch")]
+        if args.watch {
+            if let Err(e) = crate::file_watcher::start_file_watcher(kanata_arc.clone()) {
+                log::error!("Failed to start file watcher: {}", e);
+            }
         }
 
         #[cfg(target_os = "linux")]
