@@ -17,7 +17,7 @@ pub struct InputEvent(pub Stroke);
 use std::fmt;
 impl fmt::Display for InputEvent {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self)
+        write!(f, "{self:?}")
     }
 }
 
