@@ -12,6 +12,21 @@ pub enum ServerMessage {
     Error { msg: String },
 }
 
+#[derive(Serialize, Deserialize, Debug)]
+#[serde(tag = "status")]
+pub enum ServerResponse {
+    Ok,
+    Error { msg: String },
+}
+
+impl ServerResponse {
+    pub fn as_bytes(&self) -> Vec<u8> {
+        let mut msg = serde_json::to_vec(self).expect("ServerResponse should serialize");
+        msg.push(b'\n');
+        msg
+    }
+}
+
 impl ServerMessage {
     pub fn as_bytes(&self) -> Vec<u8> {
         let mut msg = serde_json::to_vec(self).expect("ServerMessage should serialize");
@@ -20,7 +35,7 @@ impl ServerMessage {
     }
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub enum ClientMessage {
     ChangeLayer {
         new: String,
@@ -35,6 +50,15 @@ pub enum ClientMessage {
     SetMouse {
         x: u16,
         y: u16,
+    },
+    Reload {},
+    ReloadNext {},
+    ReloadPrev {},
+    ReloadNum {
+        index: usize,
+    },
+    ReloadFile {
+        path: String,
     },
 }
 
@@ -51,5 +75,43 @@ impl FromStr for ClientMessage {
 
     fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
         serde_json::from_str(s)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_server_response_json_format() {
+        // Test that our API contract matches expected JSON structure
+        assert_eq!(
+            serde_json::to_string(&ServerResponse::Ok).unwrap(),
+            r#"{"status":"Ok"}"#
+        );
+        assert_eq!(
+            serde_json::to_string(&ServerResponse::Error {
+                msg: "test".to_string()
+            })
+            .unwrap(),
+            r#"{"status":"Error","msg":"test"}"#
+        );
+    }
+
+    #[test]
+    fn test_as_bytes_includes_newline() {
+        // Test our specific logic that adds newline termination
+        let response = ServerResponse::Ok;
+        let bytes = response.as_bytes();
+        assert!(bytes.ends_with(b"\n"), "Response should end with newline");
+
+        let error_response = ServerResponse::Error {
+            msg: "test".to_string(),
+        };
+        let error_bytes = error_response.as_bytes();
+        assert!(
+            error_bytes.ends_with(b"\n"),
+            "Error response should end with newline"
+        );
     }
 }
